+++ README.md
@@ -102,9 +102,8 @@
   - [x] (프로덕트 코드) `input-form` 태그 삭제 -> 테스트 실패
   - [x] (테스트 실패를 성공시키기 위해) js를 이용해 동적으로 `input-form` 태그 제작 -> 테스트 성공
   - [x] `input-form` 태그를 컴포넌트로 제작
-  - [ ] `input-form` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
-    - [ ] 동적 인자 ->`input`에 들어갈 타입 및 placeholder / label에 들어갈 이름
-    - [ ] `input-form` 테스트 코드 수정
+  - [x] `input-form` 컴포넌트 리팩토링 (동적 인자 전달/컴포넌트 이동 및 분리)
+    - [x] 동적 인자 ->`input`에 들어갈 타입 및 placeholder / label에 들어갈 이름
 
 - [ ] `textarea-form` (label + textarea.name + notice -> 묶고 있는 div)
