+++ src/main.js
@@ -1,20 +1,10 @@
 import FormItem from "./components/Form/FormItem.js";
+import InputForm from "./components/Form/InputForm.js";
 import SelectForm from "./components/Form/SelectForm.js";
 import Header from "./components/Header.js";
 import List from "./components/List.js";
 import ListItem from "./components/ListItem.js";
 import { HEADER_CONTENTS, LIST_ITEM_CONTENTS } from "./contants.js";
-
-function InputForm(type, name, required = false) {
-  const inputElement = document.createElement("input");
-
-  inputElement.type = type;
-  inputElement.name = name;
-  inputElement.id = name;
-  inputElement.required = required;
-
-  return inputElement;
-}
 
 function renderContents() {
   const app = document.getElementById("app");
